Fixed off-by-one error
There was a bug where on the second screen, if you press 1,
it shows the option on 2, press 2, show 3, etc. and if you press 9,
there's an error and it just closed. I hope I fixed that issue.
Check it out.

TL;DR: It's an off-by-one error in the second screen.